--- src/main/java/io/hhplus/tdd/point/TransactionType.java
@@ -1,11 +1,0 @@
-package io.hhplus.tdd.point;
-
-
-/**
- * 포인트 트랜잭션 종류
- * - CHARGE : 충전
- * - USE : 사용
- */
-public enum TransactionType {
-    CHARGE, USE
-}
